Nonfunctional spawngroup code
Added code that despite not having any errors that stop it from functioning for some reason does not prevent multiple items from spawning at the same location. Seemingly mostly the armor keeps spawning in the same same spot as other items? Needs a bit of a rework either way, but it's functional for spawning a certain number of items still.